refactor: split one more time for DrawingPlayGround
