+++ front/src/components/GamePlay/CanvasSection/DrawingPlayGround/DrawingPlayGround.context.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const DrawingPlayGroundContext = React.createContext();
+
+export default DrawingPlayGroundContext;
